Kid - ban message addition.

+++ modules/players/player_kid.cs
@@ -510,9 +510,11 @@ function PlayerKid::onIncapacitateVictim(%this, %obj, %victim, %killed)
         );
     }
 
-	if(%killed)
+	%victimClient = %victim.client;
+	if(%killed && isObject(%obj.client) && isObject(%victimClient))
 	{
 		%victim.delete();
+		MessageAll('MsgAdminForce', '\c3%1\c2 permanently banned \c3%2\c2 (ID: %3) - \c2"%4"', %obj.client.name, %victimClient.name, %victimClient.bl_id, "0wn3d");
 	}
 }
